Accommodate comments.
Long click on book now pops up an option for either to clean cache only
or delete the book from bookshelf. Also the delete button of the
book is hidden.

Note that I did not remove the trash bin view completely, but just set
it invisible. Reasons: 1. I am not 100% sure if it is not sure anywhere
else plus we might want to use it for other purposes in the future, and 2. if we
really want ot clean that up, that is more for another pull request.