fixes #20 Support search in bookshelf
Note that the current implementation simply search on the web first and
filter out any books that's not on the bookshelf, becuase we don't store
any information of books that's not cached other than an integer aid.
Please comment if this is not the case or you have a better idea of doing this.